// src/models/propertyModel/propertyFavoriteModel.tsx
export interface PropertyFavoriteModel {
  id: string;
  bathrooms: number;
  bedrooms: number;
  createdBy: string;
  imageThumbnailUrl: string;
  isFavourite: boolean;
  location: any;
  priceUsd: number;
  priceVnd: number;
  priorityStatus: number;
  size: number;
  title: number;
  totalImage: number;
  type: number;
  userInfo: any;
}